Expand stream slot ids to 32 bit
In a load testing scenario, we observed that an overflow in the
stream slot ids could lead to a deadlocked VAST binary which
was spinning at 100% cpu while unable to process any input.

This overflow would happen after around 1 day of heavy-duty
usage of VAST.

This change is a workaround to buy more time until the overflow
occurs; until we can find a permanent solution to the issue.