Fix container overflow reported by ASan
This fixes an ASan container overflow for macOS in Debug builds.
We ran into this previously for the unit test suite only, but after
updating my OS I now get this even for VAST itself in Debug builds. This
fixes the container overflow, but admittedly I don't understand why.
Maybe it's just a bug in ASan. :shrug: